+++ lib.php
@@ -56,13 +56,14 @@ function local_navbarplus_render_navbar_output() {
             $itemtitle = null;
             $itemvisible = false;
             $itemopeninnewwindow = false;
+            $itemadditionalclasses = null;
 
             // Make a new array on delimiter "|".
             $settings = explode('|', $line);
             // Check for the mandatory conditions first.
             // If array contains too less or too many items, do not proceed and therefore do not display the item.
             // Furthermore check it at least the first three mandatory params are not an empty string.
-            if (count($settings) >= 3 && count($settings) <= 5 &&
+            if (count($settings) >= 3 && count($settings) <= 6 &&
                 $settings[0] !== '' && $settings[1] !== '' && $settings[2] !== '') {
                 foreach ($settings as $i => $setting) {
                     $setting = trim($setting);
@@ -113,6 +114,10 @@ function local_navbarplus_render_navbar_output() {
                                     $itemvisible = true;
                                 }
                                 break;
+                            // Check for optional sixth parameter: additional classes.
+                            case 5:
+                                $itemadditionalclasses = $setting;
+                                break;
                         }
                     }
                 }
@@ -126,8 +131,14 @@ function local_navbarplus_render_navbar_output() {
                 if ($itemopeninnewwindow) {
                     $attributes['target'] = '_blank';
                 }
+                // Define classes for all icons.
+                $itemclasses = 'localnavbarplus nav-link';
+                // Add optional individual classes.
+                if (!empty($itemadditionalclasses)) {
+                    $itemclasses .= ' ' . $itemadditionalclasses;
+                }
                 // Add the link to the HTML.
-                $output .= html_writer::start_tag('div', array('class' => 'localnavbarplus nav-link'));
+                $output .= html_writer::start_tag('div', array('class' => $itemclasses));
                 $output .= html_writer::link($itemurl, $itemicon, $attributes);
                 $output .= html_writer::end_tag('div');
             }